# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RuiQi 应用防火墙管理系统 API",
        "title": "RuiQi-WAF API",
        "termsOfService": "https://github.com/HUAHUAI23/RuiQi",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/HUAHUAI23/RuiQi",
            "email": "huahua1319873800@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:2333",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/blocked-ips": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取被封禁的IP地址列表，支持分页、过滤和排序",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "封禁IP管理"
                ],
                "summary": "获取封禁IP列表",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "页码，从1开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量，最大100",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "192.168.1.1",
                        "description": "IP地址过滤，支持模糊匹配",
                        "name": "ip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "high_frequency_attack",
                        "description": "封禁原因过滤",
                        "name": "reason",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "active",
                            "expired",
                            "all"
                        ],
                        "type": "string",
                        "default": "all",
                        "description": "状态过滤：active-生效中，expired-已过期，all-全部",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "blocked_at",
                            "blocked_until",
                            "ip"
                        ],
                        "type": "string",
                        "default": "blocked_at",
                        "description": "排序字段",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "排序方向：asc-升序，desc-降序",
                        "name": "sortDir",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取封禁IP列表成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.BlockedIPListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "403": {
                        "description": "禁止访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            }
        },
        "/api/v1/blocked-ips/cleanup": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除已过期的封禁IP记录，释放存储空间",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "封禁IP管理"
                ],
                "summary": "清理过期的封禁IP记录",
                "responses": {
                    "200": {
                        "description": "清理完成",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.BlockedIPCleanupResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "403": {
                        "description": "禁止访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            }
        },
        "/api/v1/blocked-ips/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取封禁IP的统计信息，包括总数、生效数、过期数、按原因统计和按小时统计",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "封禁IP管理"
                ],
                "summary": "获取封禁IP统计信息",
                "responses": {
                    "200": {
                        "description": "获取统计信息成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.BlockedIPStatsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "403": {
                        "description": "禁止访问",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/model.ErrResponseDontShowError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BlockedIPCleanupResponse": {
            "description": "清理过期封禁IP记录的响应数据",
            "type": "object",
            "properties": {
                "deletedCount": {
                    "description": "删除的记录数量",
                    "type": "integer",
                    "example": 25
                },
                "message": {
                    "description": "操作结果消息",
                    "type": "string",
                    "example": "已成功清理过期的封禁IP记录"
                }
            }
        },
        "dto.BlockedIPHourlyStats": {
            "description": "按小时的封禁统计",
            "type": "object",
            "properties": {
                "count": {
                    "description": "该小时的封禁数量",
                    "type": "integer",
                    "example": 10
                },
                "hour": {
                    "description": "小时时间点",
                    "type": "string",
                    "example": "2023-12-01T10:00:00Z"
                }
            }
        },
        "dto.BlockedIPListResponse": {
            "description": "封禁IP分页列表响应",
            "type": "object",
            "properties": {
                "items": {
                    "description": "IP列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.BlockedIPResponse"
                    }
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer",
                    "example": 1
                },
                "pages": {
                    "description": "总页数",
                    "type": "integer",
                    "example": 10
                },
                "size": {
                    "description": "每页数量",
                    "type": "integer",
                    "example": 10
                },
                "total": {
                    "description": "总数量",
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "dto.BlockedIPResponse": {
            "description": "封禁IP详细信息",
            "type": "object",
            "properties": {
                "blockedAt": {
                    "description": "封禁开始时间",
                    "type": "string",
                    "example": "2023-12-01T10:00:00Z"
                },
                "blockedUntil": {
                    "description": "封禁结束时间",
                    "type": "string",
                    "example": "2023-12-01T11:00:00Z"
                },
                "ip": {
                    "description": "被封禁的IP地址",
                    "type": "string",
                    "example": "192.168.1.1"
                },
                "isActive": {
                    "description": "是否仍在封禁中",
                    "type": "boolean",
                    "example": true
                },
                "reason": {
                    "description": "封禁原因",
                    "type": "string",
                    "example": "high_frequency_attack"
                },
                "remainingTTL": {
                    "description": "剩余封禁时间（秒）",
                    "type": "integer",
                    "example": 3600
                },
                "requestUri": {
                    "description": "请求URI",
                    "type": "string",
                    "example": "/api/v1/login"
                }
            }
        },
        "dto.BlockedIPStatsResponse": {
            "description": "封禁IP统计信息",
            "type": "object",
            "properties": {
                "activeBlocked": {
                    "description": "当前生效的封禁数量",
                    "type": "integer",
                    "example": 50
                },
                "expiredBlocked": {
                    "description": "已过期的封禁数量",
                    "type": "integer",
                    "example": 950
                },
                "last24HourStats": {
                    "description": "最近24小时统计",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.BlockedIPHourlyStats"
                    }
                },
                "reasonStats": {
                    "description": "按原因统计",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "totalBlocked": {
                    "description": "总封禁数量",
                    "type": "integer",
                    "example": 1000
                }
            }
        },
...
        "model.ErrResponse": {
            "description": "错误的API响应标准格式",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "参数错误"
                },
                "message": {
                    "type": "string",
                    "example": "请求参数错误"
                },
                "requestId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
        "model.ErrResponseDontShowError": {
            "description": "错误的API响应标准格式,不展示 error",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "请求参数错误"
                },
                "requestId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
...
        "model.SuccessResponse": {
            "description": "成功的API响应标准格式",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "操作成功"
                },
                "requestId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "使用 Bearer {token} 格式进行身份验证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}